backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dwave-neal>=0.6.0",
    "dwave-ocean-sdk>=9.1.0",
    "fastapi>=0.123.0",
    "jijmodeling>=1.14.1",
    "openjij>=0.11.6",
    "pydantic>=2.12.5",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.7",
]
